Added Message Pushing Logic in Messages.jsx
A database scheme must be designed that will contain all the message data being sent per ticket.

Messages sent can be stored in the database with a limit of 500 characters/reply.

Messages should be identifiable by its sender, ticket-number, content, and timestamp.

Create a function in the backend that allows this message storing logic.